agents/security-specialist/requirements.txt
# GCP Libraries

google-cloud-firestore
google-cloud-aiplatform

# Git operations

GitPython